pip-requirements.txt
absl-py==0.9.0
astor==0.8.1
backcall==0.1.0
catkin-pkg==0.4.16
catkin-tools==0.4.5
cycler==0.10.0
decorator==4.4.1
distro==1.4.0
docutils==0.16
empy==3.3.4
future==0.18.2
gast==0.3.3
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.16.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.0.3
numpy==1.18.1
opencv-python==4.1.2.30
osrf-pycommon==0.1.9
parso==0.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.2.2
progress==1.5
prompt-toolkit==2.0.10
protobuf==3.11.2
ptyprocess==0.6.0
Pygments==2.5.2
pyparsing==2.4.6
python-dateutil==2.8.1
PyYAML==5.3
rosdistro==0.8.0
rospkg==1.2.3
scipy==1.4.1
six==1.14.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
torch==1.3.1
torchvision==0.4.2
traitlets==4.3.3
trollius==2.1.post2
wcwidth==0.1.8
Werkzeug==0.16.1
wrapt==1.11.2
